requirements.txt
Django==2.0.5
Pillow==5.1.0
django_cron==0.5.0
